V3: File re-organization i.e. pages & components
