+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-app",
-  "version": "2.0.0",
+  "version": "3.0.0",
   "private": true,
   "dependencies": {
     "@material-ui/core": "^4.10.2",
