--- src/components/Action/index.js
@@ -1,5 +1,0 @@
-import RecoverEmail from './RecoverEmail';
-import ResetPassword from './ResetPassword';
-import VerifyEmail from './VerifyEmail';
-
-export { RecoverEmail, ResetPassword, VerifyEmail };
