--- src/components/PasswordForget/index.js
@@ -1,4 +1,0 @@
-import PasswordForgetForm from './passwordForgetForm';
-import PasswordForgetLink from './passwordForgetLink';
-
-export { PasswordForgetForm, PasswordForgetLink };
