--- src/components/SignIn/index.js
@@ -1,4 +1,0 @@
-import SignInForm from './signInForm';
-import SignInLink from './signInLink';
-
-export { SignInForm, SignInLink };
