--- src/components/SignUp/index.js
@@ -1,4 +1,0 @@
-import SignUpForm from './signUpForm';
-import SignUpLink from './signUpLink';
-
-export { SignUpForm, SignUpLink };
