+++ src/constants/routes.js
@@ -1,8 +1,8 @@
 export const ACCOUNT = '/account';
-export const ACCOUNT_MANAGE = '/account/manage'
 export const ACTION = '/action';
 export const HOME = '/home';
 export const LANDING = '/';
 export const PASSWORD_FORGET = '/pw-forget';
+export const SETTINGS = '/settings';
 export const SIGN_IN = '/sign-in';
 export const SIGN_UP = '/sign-up';
