+++ src/pages/Action/VerifyEmail/index.js
