+++ src/pages/PasswordForget/passwordForgetLink.js
@@ -2,7 +2,6 @@ import React, { Component } from 'react';
 import { Link as RouterLink } from 'react-router-dom';
 
 import Link from '@material-ui/core/Link';
-
 import Typography from '@material-ui/core/Typography';
 
 import * as ROUTES from '../../constants/routes';
