+++ src/pages/Settings/index.js
@@ -1,20 +1,18 @@
 import React, { Component } from 'react';
-import { Link as RouterLink } from 'react-router-dom';
 import { compose } from 'recompose';
 
-import { UpdateEmail, UpdatePassword, UpdateProfile } from '../../components/Account';
+import UpdateEmail from './UpdateEmail';
+import UpdatePassword from './UpdatePassword';
+import UpdateProfile from './UpdateProfile';
 
 import Box from '@material-ui/core/Box';
-import Button from '@material-ui/core/Button';
 import Container from '@material-ui/core/Container';
 import Paper from '@material-ui/core/Paper';
 import Typography from '@material-ui/core/Typography';
 
 import { withAuthorization, withEmailVerification } from '../../session';
 
-import * as ROUTES from '../../constants/routes';
-
-class AccountManageBase extends Component {
+class SettingsBase extends Component {
   render() {
     return(
       <React.Fragment>
@@ -23,7 +21,7 @@ class AccountManageBase extends Component {
             <Paper elevation={0}>
               <Box p={3}>
                 <Typography align="center" variant="h4" gutterBottom>
-                  <strong>Manage Account</strong>
+                  <strong>Settings</strong>
                 </Typography>
                 <Typography align="center" variant="body2" gutterBottom>
                   This page is only accessible to logged in users.
@@ -43,16 +41,6 @@ class AccountManageBase extends Component {
           <Box pt={2}>
             <UpdatePassword />
           </Box>
-
-          <Box pt={2}>
-            <Paper elevation={0}>
-              <Box p={3}>
-                <Button fullWidth size="large" color="primary" component={RouterLink} to={ROUTES.ACCOUNT}>
-                  View Account
-                </Button>
-              </Box>
-            </Paper>
-          </Box>
         </Container>
       </React.Fragment>
     );
@@ -61,9 +49,11 @@ class AccountManageBase extends Component {
 
 const condition = authUser => !!authUser;
 
-const AccountManage = compose(
+const Settings = compose(
   withAuthorization(condition),
   withEmailVerification,
-)(AccountManageBase);
+)(SettingsBase);
 
-export default AccountManage;
+export default Settings;
+
+export { UpdateEmail, UpdatePassword, UpdateProfile };
